+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2026-01-18
+### Added:
+- Tree: show/print_tree method to support rich format.
+### Fixed:
+- Tests: Test not showing up correctly with split instead of splitlines.
+
 ## [1.1.0] - 2026-01-01
 ### Added:
 - Tree: Added get_subtree method as plugin.